Increase URL loading timeout in WebView
The timeout value for URL loading in the WebView component has been increased. The previous timeout of 250ms was too short and could potentially lead to problems with loading more complex webpages. It has now been extended to 10,000ms to ensure that all pages have enough time to fully load.